Add more exports
